Learning Assistant/requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
chromadb>=0.5.5
sentence-transformers>=3.0.1
rank-bm25>=0.2.2
pypdf>=4.3.1
streamlit>=1.37.1
python-dotenv>=1.0.1
pytesseract>=0.3.10
pillow>=10.4.0
